backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
celery==5.4.0
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
django==5.1.5
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
    #   django-soft-delete
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-celery-beat==2.7.0
    # via -r requirements.in
django-celery-results==2.5.1
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-soft-delete==1.0.18
    # via -r requirements.in
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-api-key==3.0.0
    # via -r requirements.in
djangorestframework-simplejwt==5.4.0
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
packaging==24.2
    # via djangorestframework-api-key
prompt-toolkit==3.0.50
    # via click-repl
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
pyyaml==6.0.2
    # via drf-spectacular
redis==5.2.1
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
tzdata==2025.1
    # via
    #   celery
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-spectacular
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
